Refactor for async storage
