+++ examples/ml/singletons.py
@@ -7,7 +7,7 @@
 from pathlib import Path
 
 from alsek import Broker
-from alsek.storage.backends.disk import DiskCacheBackend
+from alsek.storage.backends.disk.synchronous import DiskCacheBackend
 from alsek.storage.result import ResultStore
 
 cache_dir = Path("~/alsek/diskcache").expanduser()
